util methods; fixed a bug in string handling
Created methods to print tokens and to check if a token isEOF;
created a method to read a list of tokens from a src until EOF
or a newline.

Started thinking about how the prompt-interactive environment
may work.

(Probably should create a method to read from a src until EOF,
as it may be more useful to handle files).

Fixed a bug in string handling: a string should always end
in the same line that it begins in (it is not a string block).
Also it should change this.hadError to true and add the error
to the errorList.