Merged some cases in the switch statement. Fixed handleAlphaToken (it did not read digits)
Renamed and repurposed readUntilNewLineOrEOF. It is now just used to read inputs from the REPL. It does not append EOF to each new line.

readUntilEOF should be used to parse files

Main now reads lines and prints the token result until EOF is given as input.

Various small fixes and cleanup of code