First version of the Interpreter class && minor fixes
Right now the REPL is able to parse and evaluate expressions, working
as a simple calculator.
I have yet to handle Statements, and, to clean some things (and maybe
rewrite others), but it is working nicely.
Added first simple tests for the expression evaluation.

Added line position to the Token types, so that debugging is easier for a
hypothetical user.