implemented break && continue; some fixes; test
Break and continue are handled as throwables that are catched if they are inside a loop.
If they are not inside a loop, they raise an exception.
This is probably not the best way to do this, but it is the most straightforward
way I have found.

Previously the lexer was not consuming some chars that it should have
The astprinter was printing some objects pointers instead of valid text
The if statement now consumes the left bracket and returns a block
statement (previously it could not handle blocks, since the LBrace was already
consumed)

The while loop is a new kind of statement that executes until condition becomes
true or is exited; the for loop is a wrapper around this while statement.
It moves the initializer above the while statement, the condition inside the while
and the increment at the end of the while loop.

New tests.